remove profile_t from master_t.
this also fixes unnecessary offset from rec->ptr in cliPrintVar and
cliSetVar - rec->ptr is already correct at the time the code as called.
the intention was to use rec->base + an offset but that's not needed.

also fixes incorrect size for pgn in configRecord_t.

it's interesting how this is a major change to the structure but with
very little code impact.